Seperated Invoice, OrderDetail, OrderUpdate, Logout, ResellerPhone

+++ app/Http/Controllers/ShopController.php
@@ -38,41 +38,79 @@ class ShopController extends Controller
      */
     public function store(Request $request)
     {
-        $data = $request->validate([
+        tap($request->validate([
             'name' => 'required|max:255',
             'email' => 'required|email',
             'phone' => 'required',
             'logo' => 'nullable|image|mimes:jpeg,png,jpg,gif,svg|max:2048',
         ]) + [
             'reseller_id' => auth('reseller')->user()->id
-        ];
+        ], function($data) use ($request) {
+            
+            if($request->hasFile('logo')) {
+                $filenameWithExt = $request->file('logo')->getClientOriginalName(); 
+                $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME); 
+                $extension = $request->file('logo')->getClientOriginalExtension(); 
+                $fileNameToStore= $filename.time().'.'.$extension; 
+                $thumbnailpic= 'thumb'.'-'.$fileNameToStore;
 
-        $filenameWithExt = $request->file('logo')->getClientOriginalName(); 
-        $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME); 
-        $extension = $request->file('logo')->getClientOriginalExtension(); 
-        $fileNameToStore= $filename.time().'.'.$extension; 
-        $thumbnailpic= 'thumb'.'-'.$fileNameToStore;
+                //This store image creates the folder and saves the file 
+                $path = $request->file('logo')->storeAs('public/shop', $fileNameToStore);
 
-        //This store image creates the folder and saves the file 
-        $path = $request->file('logo')->storeAs('public/shop', $fileNameToStore);
+                if(! is_dir($dir = public_path('shop'))) {
+                    mkdir($dir, 755);
+                }
+                $to = 'shop/' . $thumbnailpic;
 
-        if(! is_dir($dir = public_path('shop'))) {
-            mkdir($dir, 755);
-        }
-        $to = 'shop/' . $thumbnailpic;
+                //Here is where I am trying to resize with image and it breaks
+                Image::make( storage_path().'/app/public/shop/'.$fileNameToStore)->resize(250, 66)->save(public_path($to));
 
-        //Here is where I am trying to resize with image and it breaks
-        Image::make( storage_path().'/app/public/shop/'.$fileNameToStore)->resize(250, 66)->save(public_path($to));
+                if(!empty($to)) {
+                    Storage::delete(public_path($to));
+                    unset($data['logo']);
+                    $data += [
+                        'logo' => $to
+                    ];
+                }
+            }
+            
+            $shop = Shop::create($data);
+        });
+        // $data = $request->validate([
+        //     'name' => 'required|max:255',
+        //     'email' => 'required|email',
+        //     'phone' => 'required',
+        //     'logo' => 'nullable|image|mimes:jpeg,png,jpg,gif,svg|max:2048',
+        // ]) + [
+        //     'reseller_id' => auth('reseller')->user()->id
+        // ];
 
-        if(!empty($to)) {
-            Storage::delete(public_path($to));
-            unset($data['logo']);
-            $data += [
-                'logo' => $to
-            ];
-        }
+        // $filenameWithExt = $request->file('logo')->getClientOriginalName(); 
+        // $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME); 
+        // $extension = $request->file('logo')->getClientOriginalExtension(); 
+        // $fileNameToStore= $filename.time().'.'.$extension; 
+        // $thumbnailpic= 'thumb'.'-'.$fileNameToStore;
+
+        // //This store image creates the folder and saves the file 
+        // $path = $request->file('logo')->storeAs('public/shop', $fileNameToStore);
+
+        // if(! is_dir($dir = public_path('shop'))) {
+        //     mkdir($dir, 755);
+        // }
+        // $to = 'shop/' . $thumbnailpic;
+
+        // //Here is where I am trying to resize with image and it breaks
+        // Image::make( storage_path().'/app/public/shop/'.$fileNameToStore)->resize(250, 66)->save(public_path($to));
+
+        // if(!empty($to)) {
+        //     Storage::delete(public_path($to));
+        //     unset($data['logo']);
+        //     $data += [
+        //         'logo' => $to
+        //     ];
+        // }
         
-        $shop = Shop::create($data);
+        // $shop = Shop::create($data);
 
         return redirect()->route('reseller.shops.index')->with('success', 'Shop Has Created Successfully.');
     }
